requirements.txt
Django==4.2
psycopg2>=2.7,<3.0
django-phone-field
phonenumbers
django-phonenumber-field